[Refactor] ai 폴더 budget, type 폴더로 세부 분리
