+++ ai/type/utils.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pandas as pd
 import ast
 from scipy.stats import entropy
